Fixed client - socket communication

+++ Client/Player.cpp
@@ -1,6 +1,7 @@
 #include "Player.hpp"
 #include "Constants.hpp"
 #include "Utils.hpp"
+#include <iostream>
 
 
 void Player::update(Camera& camera, Settings& settings, std::vector<vec3>& colors, ParticleRenderer& renderer, World& world, float deltaTime, sf::TcpSocket& socket){
@@ -45,7 +46,7 @@ void Player::placeBlocks(Camera& camera, const std::vector<vec3>& colors, Partic
 
                     //Sending block update information to server
                     sf::Packet packet;
-                    packet << (uint8_t)rayPosition.x << (uint8_t)rayPosition.y << (uint8_t)rayPosition.z << selectedBlock;
+                    packet << (int)rayPosition.x << (int)rayPosition.y << (int)rayPosition.z << (uint8_t)selectedBlock;
                     socket.send(packet);
                     world.setBlock(rayPosition.x, rayPosition.y, rayPosition.z, selectedBlock);
